requirements.txt
python-telegram-bot[job-queue]
openai
python-dotenv
requests
beautifulsoup4
